Fix: Camera Jitter
The players animation would causes the camera for move plus and negative 0.03, this meant that when calcuating angles based on user position the angle was always different even when moving forward because of this slight change

New Camera Reference is used which contains Transform snapping. This locks the transform from moving this small amount each time.